Access control for participant edit

+++ src/Controller/ParticipantController.php
@@ -57,48 +57,53 @@ class ParticipantController extends AbstractController
     #[Route('/{id}/edit', name: 'app_participant_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Participant $participant, ImageRepository $imageRepository, ParticipantRepository $participantRepository, UserRepository $userRepository,SluggerInterface $slugger, EntityManagerInterface $entityManager): Response
     {
-        // TODO : Controle utilisateur
-        $form = $this->createForm(ParticipantType::class, $participant);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $imageFile = $form->get('image')->getData();
-            $image = new Image();
-            if ($imageFile) {
+        $loggedUser = $this->getUser();
+        if(in_array('ROLE_ADMIN',$loggedUser->getRoles()) || $loggedUser->getParticipant()->getId()===$participant->getId()){
+            $form = $this->createForm(ParticipantType::class, $participant);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $imageFile = $form->get('image')->getData();
+                $image = new Image();
+                if ($imageFile) {
 
-                $previousImage = $participant->getImage();
-                if (isset($previousImage)) {
-                    $imageRepository->remove($previousImage,true);
-                }
-                $originalFilename = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
-                // On viens nettoyer le nom du fichier pour éviter tout problème dans l'URL
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename.'-'.uniqid().'.'.$imageFile->guessExtension();
+                    $previousImage = $participant->getImage();
+                    if (isset($previousImage)) {
+                        $imageRepository->remove($previousImage,true);
+                    }
+                    $originalFilename = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
+                    // On viens nettoyer le nom du fichier pour éviter tout problème dans l'URL
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename.'-'.uniqid().'.'.$imageFile->guessExtension();
 
-                // On essaye de déplacer le fichier dans le dossier "profile_images"
-                try {
-                    $profileImagesDirectory = $this->getParameter('profile_images');
-                    $imageFile->move($profileImagesDirectory, $newFilename);
-                } catch (FileException $e) {
-                    // ... On gère l'exception
+                    // On essaye de déplacer le fichier dans le dossier "profile_images"
+                    try {
+                        $profileImagesDirectory = $this->getParameter('profile_images');
+                        $imageFile->move($profileImagesDirectory, $newFilename);
+                    } catch (FileException $e) {
+                        // ... On gère l'exception
+                    }
+                    // On enregistre le nom du fichier plutôt que le fichier lui même
+                    $image->setImageFile($newFilename);
                 }
-                // On enregistre le nom du fichier plutôt que le fichier lui même
-                $image->setImageFile($newFilename);
+
+                $user = $participant->getUser();
+                $participant->setImage($image);
+                $user->setEmail($participant->getMail());
+                $user->setUsername($form->get("username")->getData());
+                $userRepository->save($user, true);
+                $participantRepository->save($participant, true);
+
+                return $this->redirectToRoute('app_participant_index', [], Response::HTTP_SEE_OTHER);
             }
 
-            $user = $participant->getUser();
-            $participant->setImage($image);
-            $user->setEmail($participant->getMail());
-            $user->setUsername($form->get("username")->getData());
-            $userRepository->save($user, true);
-            $participantRepository->save($participant, true);
-
-            return $this->redirectToRoute('app_participant_index', [], Response::HTTP_SEE_OTHER);
+            return $this->render('participant/edit.html.twig', [
+                'participant' => $participant,
+                'form' => $form,
+            ]);
+        }else {
+            $this->addFlash("danger","Vous n'avez pas les droits.");
+            return $this->redirectToRoute("activity_list");
         }
-
-        return $this->render('participant/edit.html.twig', [
-            'participant' => $participant,
-            'form' => $form,
-        ]);
     }
 
     #[Route('/admin/{id}', name: 'app_participant_delete', methods: ['POST'])]
